sincronizacion-obsidiane/requirements_dualcore.txt
# requirements_dualcore.txt
# Dependencias optimizadas para AMD Dual-Core + 8GB RAM
# Versiones FIJAS para reproducibilidad

# ============================================================
# Core Científico
# ============================================================
numpy==1.24.3              # Computación numérica (soporte fundamental)
scipy==1.11.2              # Algoritmos científicos
scikit-learn==1.3.0        # Machine learning ligero
pandas==2.0.3              # Manejo de datos tabular

# ============================================================
# NLP Y EMBEDDINGS (CRÍTICO PARA MEMORIA)
# ============================================================
# IMPORTANTE: Usar modelos "sm" (small), NUNCA "lg" (large)
spacy[cuda-auto]==3.7.2    # NLP - se descarga modelo aparte
sentence-transformers==2.2.2  # Embeddings (all-MiniLM-L6-v2 = 80MB)
transformers==4.34.0       # Base para sentence-transformers
torch==2.1.0               # PyTorch CPU

# Instalación de modelo spaCy pequeño:
# python -m spacy download es_core_news_sm

# ============================================================
# GRAFOS E ALGORITMOS
# ============================================================
networkx==3.1              # Algoritmos de grafos (local, NetworkX)
neo4j==5.12.0              # Conexión Neo4j Bolt

# ============================================================
# CLUSTERING Y DIMENSIONALIDAD
# ============================================================
scikit-optimize==0.9.1     # Optimización Bayesiana
umap-learn==0.5.3          # UMAP (reducción dimensional)
hdbscan==0.8.33            # Clustering robusto

# ============================================================
# CONFIGURACIÓN Y UTILITIES
# ============================================================
pyyaml==6.0                # Cargar config YAML
python-dotenv==1.0.0       # Variables de entorno
psutil==5.9.5              # Monitoreo de sistema
requests==2.31.0           # HTTP client

# ============================================================
# LOGGING Y DIAGNÓSTICOS
# ============================================================
loguru==0.7.2              # Logging mejorado
colorama==0.4.6            # Colores en terminal
tqdm==4.66.1               # Barras de progreso

# ============================================================
# LIGHTRAG (Desde GitHub - instalar después)
# ============================================================
# pip install git+https://github.com/HKUDS/LightRAG.git

# ============================================================
# DESARROLLO (Opcional)
# ============================================================
# pytest==7.4.2
# black==23.9.1
# pylint==2.18.0
